+++ src/javascript/SelectorTypes/Picker/Picker.constants.js
