+++ src/javascript/SelectorTypes/Picker/Picker.redux.js
@@ -1,7 +1,7 @@
 import {createActions, handleActions} from 'redux-actions';
 import {COMBINED_REDUCERS_NAME} from '~/registerReducer';
-import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
-import {toArray} from './Picker2.utils';
+import {Constants} from '~/SelectorTypes/Picker/Picker.constants';
+import {toArray} from './Picker.utils';
 
 export const {
     cePickerSite,
