+++ src/javascript/SelectorTypes/Picker/Picker.scss
