+++ src/javascript/SelectorTypes/Picker/Picker.utils.js
