+++ src/javascript/SelectorTypes/Picker/PickerDialog/index.js
@@ -1,3 +1,3 @@
-export * from './PickerDialog2';
-export {GET_PICKER_NODE} from '~/SelectorTypes/Picker/PickerDialog/PickerDialog2.gql-queries';
+export * from './PickerDialog';
+export {GET_PICKER_NODE} from '~/SelectorTypes/Picker/PickerDialog/PickerDialog.gql-queries';
 export {PickerSiteSwitcher} from '~/SelectorTypes/Picker/PickerDialog/PickerSiteSwitcher';
