+++ src/javascript/SelectorTypes/Picker/configs/editorialLinkPicker/EditorialLinkContentTypeSelector.jsx
@@ -3,8 +3,8 @@ import {FolderSpecial, Page, Tab, TabItem} from '@jahia/moonstone';
 import {useDispatch, useSelector} from 'react-redux';
 import classes from './EditorialLinkContentTypeSelector.scss';
 import {useTranslation} from 'react-i18next';
-import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
-import {cePickerSetTableViewType} from '~/SelectorTypes/Picker/Picker2.redux';
+import {Constants} from '~/SelectorTypes/Picker/Picker.constants';
+import {cePickerSetTableViewType} from '~/SelectorTypes/Picker/Picker.redux';
 import {configPropType} from '~/SelectorTypes/Picker/configs/configPropType';
 
 const localStorage = window.localStorage;
