+++ src/javascript/SelectorTypes/Picker/configs/editorialLinkPicker/PickerEditorialLinkQueryHandler.jsx
@@ -1,4 +1,4 @@
-import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
+import {Constants} from '~/SelectorTypes/Picker/Picker.constants';
 import {PickerTreeQueryHandler} from '~/SelectorTypes/Picker/configs/queryHandlers';
 
 export const PickerEditorialLinkQueryHandler = {
